Check for game end, change evaluation function

+++ NegaMaxNegaScout/NegaMaxNegaScout.BLL/Entities/Position.cs
@@ -4,5 +4,5 @@ public class Position
 {
     public (int, int) PlayerPosition { get; set; }
     public (int, int) EnemyPosition { get; set; }
-    public int Evaluation { get; set; }
+    public float Evaluation { get; set; }
 }
